+++ Extensions/Minded.Extensions.WebApi/ICommandRestRule.cs
@@ -1,14 +1,10 @@
 ﻿using System;
-using System.Net;
 using Minded.Framework.CQRS.Command;
 
 namespace Minded.Extensions.WebApi
 {
-    public interface ICommandRestRule
+    public interface ICommandRestRule: IMessageRestRule
     {
-        RestOperation Operation { get; }
-        HttpStatusCode ResultStatusCode { get; }
-        ContentResponse ContentResponse { get; }
         Func<ICommandResponse, bool> RuleCondition { get; }
     }
 }