+++ Extensions/Minded.Extensions.WebApi/IMessageRestRule.cs
@@ -1,0 +1,11 @@
+﻿using System.Net;
+
+namespace Minded.Extensions.WebApi
+{
+    public interface IMessageRestRule
+    {
+        RestOperation Operation { get; }
+        HttpStatusCode ResultStatusCode { get; }
+        ContentResponse ContentResponse { get; }
+    }
+}