+++ Extensions/Minded.Extensions.WebApi/IQueryRestRule.cs
@@ -1,13 +1,9 @@
 ﻿using System;
-using System.Net;
 
 namespace Minded.Extensions.WebApi
 {
-    public interface IQueryRestRule
+    public interface IQueryRestRule: IMessageRestRule
     {
-        RestOperation Operation { get; }
-        HttpStatusCode ResultStatusCode { get; }
-        ContentResponse ContentResponse { get; }
         Func<object, bool> RuleCondition { get; }
     }
 }